tests: the behavioral transpile gates were blind to nested sub-library packages
All three transpile gates -- check-no-regression.ps1, the standalone
BehavioralRunner, and the MSTest BehavioralTestBase -- enumerated their targets
with a TOP-LEVEL-only walk of Tests\Behavioral. That silently excluded the 22
sub-library packages nested inside a test folder (IoLike\FsLike,
VersionedImport\vlib, CrossPackageArrayZeroValue\bufpkg,
GoNamespaceShadow\nsshadowlib, ...): no gate ever re-transpiled them.

Two consequences, the second the one that matters.

1. Their committed .cs froze at whatever converter last touched them by hand. 17
   files across 13 packages had drifted, spanning three separate increments: the
   `// <TypeAccessibility>` block in package_info.cs (13 packages -- all 521
   top-level packages carry it), string-literal hoisting (identlib, netlike), and
   the compound-assign result cast (vlib). Their .csproj had drifted too, still
   carrying the removed <Version> element and missing both the PackageReadmeFile
   block and the corrected `*_test.cs` remove pattern that all 521 top-level
   projects already have.

2. FALSE GREEN. A sub-library's package_info.cs is an INPUT to its parent's
   transpile: the parent reads the sibling's [assembly: GoImplement] records to
   decide whether to mint a local value adapter. Because the sub-library was never
   regenerated, a converter regression in that area could not make the parent's
   golden fail -- the parent kept reading stale-but-plausible records. That
   disarmed the ForeignValueImplementSuppression, ValueAdapterDynamicType and
   SamePackageImplementNoWitness guards specifically.

All three gates now walk recursively and DEEPEST-FIRST, so a sub-library is
regenerated before its parent consumes it; both runners' UpToDate checks consider
the nested packages too, or a stale sibling would simply be skipped again.
Enumeration goes 521 -> 543 packages.

The 13 stale packages are re-baselined in this same commit. They are generated
files, and every diff was verified as already-documented converter behavior rather
than accepted on faith: the TypeAccessibility block is present in 521/521
top-level package_info.cs, hoisting is guarded by the StringLiteralHoisting test
and appears in 248 top-level goldens, and the `x ^= (uint64)(...)` compound-assign
cast is the standard emitted form throughout src/core (crypto/cipher/gcm.cs,
crypto/des/block.cs) with same-shape precedent in the goldens at other widths.

Goldens stay top-level-only -- nested packages have no .cs.target and
UpdateTestTargets is deliberately unchanged -- so nested drift is caught by CNR's
git status, while the cross-package effect is caught by the parent's golden.

Also closes two adjacent blind spots in check-no-regression.ps1 that the above
exposed: its report and its -Revert both matched Tests\Behavioral\*.cs while
excluding only BehavioralTests, so editing the runner made CNR accuse itself of a
regression -- and -Revert would have DESTROYED that uncommitted edit.

Gates:
- check-no-regression 543/543 (365s): the only drift is the intended re-baseline.
  Every one of the 521 top-level packages re-transpiled byte-identically, so the
  stale records happened to encode the same decisions today's converter makes --
  the false-green route was latent, not actively masking a live regression.
- UpdateTestTargets --createTargetFiles: zero golden churn, zero test-class churn.
- Full behavioral suite PASS (533s): 521/521 Transpile + Compile + Target, 491/491
  Output vs `go run`, 30 skipped (no `package main`).
- Direct proof the gate now reaches nested packages: deleting
  IoLike\FsLike\info.cs and running only `--filter IoLike --phase transpile`
  regenerates it, byte-identically. Under the old walk it stayed missing.

CLAUDE.md records the new counts, the re-measured baselines (CNR 365s, suite
533s), and a concurrent-session hazard hit while gating this: cleanup preambles
that kill by process NAME reach across worktrees, so two full runs died at 124s
and 163s with exit -1 and a truncated log. That signature means "killed
externally", not a compile failure -- the same corpus then passed 521/521
untouched from a uniquely-named apphost copy.

+++ src/Tests/Behavioral/VersionedImport/vlib/vlib.cs
@@ -15,9 +15,9 @@ public static ж<PCG> NewPCG(uint64 seed) {
 }
 
 [GoRecv] public static uint64 Uint64(this ref PCG p) {
-    p.state ^= (p.state << (int)(13));
-    p.state ^= (p.state >> (int)(7));
-    p.state ^= (p.state << (int)(17));
+    p.state ^= (uint64)((p.state << (int)(13)));
+    p.state ^= (uint64)((p.state >> (int)(7)));
+    p.state ^= (uint64)((p.state << (int)(17)));
     return p.state;
 }
